Move the definition for intrinsic function to the ty module.
