+++ sway-core/src/language/ty/expression/mod.rs
@@ -1,6 +1,8 @@
 #[allow(clippy::module_inception)]
 mod expression;
 mod expression_variant;
+mod intrinsic_function;
 
 pub use expression::*;
 pub use expression_variant::*;
+pub use intrinsic_function::*;
