+++ sway-core/src/semantic_analysis/ast_node/expression/intrinsic_function.rs
@@ -1,6 +1,3 @@
-use std::fmt;
-
-use itertools::Itertools;
 use sway_ast::intrinsics::Intrinsic;
 use sway_error::error::{CompileError, Hint};
 use sway_types::Span;
@@ -10,80 +7,10 @@ use crate::{
     language::{parsed::Expression, ty},
     semantic_analysis::TypeCheckContext,
     type_system::*,
-    types::DeterministicallyAborts,
     CompileResult,
 };
 
-#[derive(Debug, Clone, PartialEq)]
-pub struct TyIntrinsicFunctionKind {
-    pub kind: Intrinsic,
-    pub arguments: Vec<ty::TyExpression>,
-    pub type_arguments: Vec<TypeArgument>,
-    pub span: Span,
-}
-
-impl CopyTypes for TyIntrinsicFunctionKind {
-    fn copy_types(&mut self, type_mapping: &TypeMapping) {
-        for arg in &mut self.arguments {
-            arg.copy_types(type_mapping);
-        }
-        for targ in &mut self.type_arguments {
-            targ.type_id.copy_types(type_mapping);
-        }
-    }
-}
-
-impl fmt::Display for TyIntrinsicFunctionKind {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let targs = self
-            .type_arguments
-            .iter()
-            .map(|targ| look_up_type_id(targ.type_id))
-            .join(", ");
-        let args = self.arguments.iter().map(|e| format!("{}", e)).join(", ");
-
-        write!(f, "{}::<{}>::({})", self.kind, targs, args)
-    }
-}
-
-impl DeterministicallyAborts for TyIntrinsicFunctionKind {
-    fn deterministically_aborts(&self) -> bool {
-        matches!(self.kind, Intrinsic::Revert)
-            || self.arguments.iter().any(|x| x.deterministically_aborts())
-    }
-}
-
-impl CollectTypesMetadata for TyIntrinsicFunctionKind {
-    fn collect_types_metadata(&self) -> CompileResult<Vec<TypeMetadata>> {
-        let mut warnings = vec![];
-        let mut errors = vec![];
-        let mut types_metadata = vec![];
-        for type_arg in self.type_arguments.iter() {
-            types_metadata.append(&mut check!(
-                type_arg.type_id.collect_types_metadata(),
-                return err(warnings, errors),
-                warnings,
-                errors
-            ));
-        }
-        for arg in self.arguments.iter() {
-            types_metadata.append(&mut check!(
-                arg.collect_types_metadata(),
-                return err(warnings, errors),
-                warnings,
-                errors
-            ));
-        }
-
-        if matches!(self.kind, Intrinsic::Log) {
-            types_metadata.push(TypeMetadata::LoggedType(self.arguments[0].return_type));
-        }
-
-        ok(types_metadata, warnings, errors)
-    }
-}
-
-impl TyIntrinsicFunctionKind {
+impl ty::TyIntrinsicFunctionKind {
     pub(crate) fn type_check(
         mut ctx: TypeCheckContext,
         kind_binding: TypeBinding<Intrinsic>,
@@ -116,7 +43,7 @@ impl TyIntrinsicFunctionKind {
                     warnings,
                     errors
                 );
-                let intrinsic_function = TyIntrinsicFunctionKind {
+                let intrinsic_function = ty::TyIntrinsicFunctionKind {
                     kind,
                     arguments: vec![exp],
                     type_arguments: vec![],
@@ -163,7 +90,7 @@ impl TyIntrinsicFunctionKind {
                     warnings,
                     errors,
                 );
-                let intrinsic_function = TyIntrinsicFunctionKind {
+                let intrinsic_function = ty::TyIntrinsicFunctionKind {
                     kind,
                     arguments: vec![],
                     type_arguments: vec![TypeArgument {
@@ -206,7 +133,7 @@ impl TyIntrinsicFunctionKind {
                     warnings,
                     errors,
                 );
-                let intrinsic_function = TyIntrinsicFunctionKind {
+                let intrinsic_function = ty::TyIntrinsicFunctionKind {
                     kind,
                     arguments: vec![],
                     type_arguments: vec![TypeArgument {
@@ -219,7 +146,7 @@ impl TyIntrinsicFunctionKind {
                 (intrinsic_function, insert_type(TypeInfo::Boolean))
             }
             Intrinsic::GetStorageKey => (
-                TyIntrinsicFunctionKind {
+                ty::TyIntrinsicFunctionKind {
                     kind,
                     arguments: vec![],
                     type_arguments: vec![],
@@ -280,7 +207,7 @@ impl TyIntrinsicFunctionKind {
                     errors
                 );
                 (
-                    TyIntrinsicFunctionKind {
+                    ty::TyIntrinsicFunctionKind {
                         kind,
                         arguments: vec![lhs, rhs],
                         type_arguments: vec![],
@@ -394,7 +321,7 @@ impl TyIntrinsicFunctionKind {
                 );
 
                 (
-                    TyIntrinsicFunctionKind {
+                    ty::TyIntrinsicFunctionKind {
                         kind,
                         arguments: vec![index, tx_field_id],
                         type_arguments: vec![TypeArgument {
@@ -444,7 +371,7 @@ impl TyIntrinsicFunctionKind {
                     return err(warnings, errors);
                 }
 
-                let intrinsic_function = TyIntrinsicFunctionKind {
+                let intrinsic_function = ty::TyIntrinsicFunctionKind {
                     kind,
                     arguments: vec![exp],
                     type_arguments: vec![],
@@ -489,7 +416,7 @@ impl TyIntrinsicFunctionKind {
                     });
                     return err(warnings, errors);
                 }
-                let intrinsic_function = TyIntrinsicFunctionKind {
+                let intrinsic_function = ty::TyIntrinsicFunctionKind {
                     kind,
                     arguments: vec![exp],
                     type_arguments: vec![],
@@ -581,7 +508,7 @@ impl TyIntrinsicFunctionKind {
                         span: span.clone(),
                     }
                 });
-                let intrinsic_function = TyIntrinsicFunctionKind {
+                let intrinsic_function = ty::TyIntrinsicFunctionKind {
                     kind,
                     arguments: vec![key_exp, val_exp],
                     type_arguments: type_argument.map_or(vec![], |ta| vec![ta]),
@@ -609,7 +536,7 @@ impl TyIntrinsicFunctionKind {
                     warnings,
                     errors
                 );
-                let intrinsic_function = TyIntrinsicFunctionKind {
+                let intrinsic_function = ty::TyIntrinsicFunctionKind {
                     kind,
                     arguments: vec![exp],
                     type_arguments: vec![],
@@ -679,7 +606,7 @@ impl TyIntrinsicFunctionKind {
                     errors
                 );
                 (
-                    TyIntrinsicFunctionKind {
+                    ty::TyIntrinsicFunctionKind {
                         kind,
                         arguments: vec![lhs, rhs],
                         type_arguments: vec![],
@@ -731,7 +658,7 @@ impl TyIntrinsicFunctionKind {
                 }
 
                 (
-                    TyIntrinsicFunctionKind {
+                    ty::TyIntrinsicFunctionKind {
                         kind,
                         arguments: vec![revert_code],
                         type_arguments: vec![],
